Add Editor Config and Git Attributes files to control line endings
The article [Solving inconsistent line endings within a Git repository](https://barryvanveen.nl/blog/52-solving-inconsistent-line-endings-within-a-git-repository) suggested adding these two files as a way of trying resolving this kind of problem. [I tried it on my copy of the project](https://github.com/NicMcPhee/Machine-shop-simulator/pull/46), and it appears to have worked nicely, so hopefully that will resolve the problem here as well.